% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/robustrolling-package.R
\docType{package}
\name{robustrolling-package}
\alias{robustrolling}
\alias{robustrolling-package}
\title{robustrolling Package}
\description{
High-performance rolling window metrics with C++17 backend.
}
\keyword{internal}
