% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rolling_metrics.R
\name{rolling_min}
\alias{rolling_min}
\title{Rolling Minimum}
\usage{
rolling_min(x, window_size)
}
\arguments{
\item{x}{A numeric vector of type double.}

\item{window_size}{Positive integer window length.}
}
\value{
A numeric vector with rolling minimum values.
}
\description{
Computes the rolling minimum over a numeric vector using
a monotonic deque.
}
\examples{
x <- as.double(c(1, 3, 2, 5, 4))
rolling_min(x, 3L)
}
